+++ src/lib.rs
@@ -60,6 +60,7 @@
 //! ```
 
 mod account_resource;
+pub mod business_upgrade;
 mod campaigns;
 mod client;
 mod contacts;
@@ -79,6 +80,7 @@ mod webhook_resource;
 pub mod webhooks;
 
 pub use account_resource::AccountResource;
+pub use business_upgrade::BusinessUpgradeResource;
 pub use campaigns::*;
 pub use client::{Sendly, SendlyConfig};
 pub use contacts::*;